mv files to src

+++ algorithmic-toolbox/week1/src/MaxPairwiseProduct.java
@@ -2,9 +2,9 @@ import java.util.*;
 import java.io.*;
 
 public class MaxPairwiseProduct {
-    static int getMaxPairwiseProduct(int[] numbers) {
-        int result = 0;
-        int n = numbers.length;
+    static long getMaxPairwiseProduct(long[] numbers) {
+        long result = 0;
+        long n = numbers.length;
         for (int i = 0; i < n; ++i) {
             for (int j = i + 1; j < n; ++j) {
                 if (numbers[i] * numbers[j] > result) {
@@ -17,10 +17,10 @@ public class MaxPairwiseProduct {
 
     public static void main(String[] args) {
         FastScanner scanner = new FastScanner(System.in);
-        int n = scanner.nextInt();
-        int[] numbers = new int[n];
+        long n = scanner.nextLong();
+        long[] numbers = new long[(int)n];
         for (int i = 0; i < n; i++) {
-            numbers[i] = scanner.nextInt();
+            numbers[i] = scanner.nextLong();
         }
         System.out.println(getMaxPairwiseProduct(numbers));
     }
@@ -48,8 +48,8 @@ public class MaxPairwiseProduct {
             return st.nextToken();
         }
 
-        int nextInt() {
-            return Integer.parseInt(next());
+        long nextLong() {
+            return Long.parseLong(next());
         }
     }
